debian/rules: apply the best practice of DEB_BUILD
By following the instruction of dpkg-buildpackage manual:

"
Even if dpkg-buildpackage exports some variables, debian/rules should
not rely on their presence and should instead use the respective
interface to retrieve the needed values, because that file is the main
entry point to build packages and running it standalone should be
supported.
"

Let us use dpkg-buildpackage to retrieve the needed value.